unit test refactor
